\name{reshape_values}
\alias{reshape_values}
\title{Reshape a Numeric Vector}
\usage{
reshape_values(x, nrow)
}
\arguments{
\item{x}{A numeric vector.}
\item{nrow}{A non-negative number of rows.}
}
\value{
A numeric matrix.
}
\description{
Returns a numeric vector with a validated two-element \code{dim} attribute.
}
